+++ changelog.md
@@ -1,5 +1,6 @@
 #### 0.0.3 (2021-02-17)
 
+- add `user_trailingslashit` for front page redirect (#cf9bd5b)
 - Support for language switcher in admin bar on frontend (#9957010)
 
 #### 0.0.2 (2021-02-17)