+++ changelog.md
@@ -1,6 +1,7 @@
-#### 0.0.3 (2021-02-17)
+#### 0.0.3 (2021-02-23)
 
-- add `user_trailingslashit` for front page redirect (#cf9bd5b)
+- Skip `pre_get_posts` if `suppress_filters==true` or query is for an attachment. (#eb95ea1)
+- add `user_trailingslashit` for front page redirect (#f25217a)
 - Support for language switcher in admin bar on frontend (#9957010)
 
 #### 0.0.2 (2021-02-17)